refactor(enchantments): merge min and max enchantment level expressions
